+++ src/features/speedrun/season3/callbacks/preItemPickup.ts
@@ -1,18 +1,7 @@
-import {
-  CollectibleType,
-  ItemType,
-  LevelStage,
-  RoomType,
-  TrapdoorVariant,
-} from "isaac-typescript-definitions";
-import {
-  isRoomInsideGrid,
-  onRepentanceStage,
-  PickingUpItem,
-  spawnTrapdoorWithVariant,
-} from "isaacscript-common";
+import { TrapdoorVariant } from "isaac-typescript-definitions";
+import { PickingUpItem, spawnTrapdoorWithVariant } from "isaacscript-common";
 import { ChallengeCustom } from "../../../../enums/ChallengeCustom";
-import g from "../../../../globals";
+import { inClearedMomBossRoom } from "../../../../utilsGlobals";
 import { season3HasDogmaGoal } from "../v";
 
 /** One tile away from the bottom door in a 1x1 room. */
@@ -20,39 +9,23 @@ export const INVERTED_TRAPDOOR_GRID_INDEX = 97;
 
 export function season3PreItemPickup(
   _player: EntityPlayer,
-  pickingUpItem: PickingUpItem,
+  _pickingUpItem: PickingUpItem,
 ): void {
   const challenge = Isaac.GetChallenge();
   if (challenge !== ChallengeCustom.SEASON_3) {
     return;
   }
 
-  spawnTrapdoorOnTakePhoto(pickingUpItem);
+  spawnTrapdoorOnTakeMomCollectible();
 }
 
-function spawnTrapdoorOnTakePhoto(pickingUpItem: PickingUpItem) {
-  if (pickingUpItem.itemType !== ItemType.PASSIVE) {
-    return;
-  }
-
-  if (
-    pickingUpItem.subType !== CollectibleType.POLAROID &&
-    pickingUpItem.subType !== CollectibleType.NEGATIVE
-  ) {
+/** The trapdoor to the Dogma goal should only spawn after the players have taken a collectible. */
+function spawnTrapdoorOnTakeMomCollectible() {
+  if (!inClearedMomBossRoom()) {
     return;
   }
 
   if (!season3HasDogmaGoal()) {
-    return;
-  }
-
-  const stage = g.l.GetStage();
-  if (stage !== LevelStage.DEPTHS_2 || onRepentanceStage()) {
-    return;
-  }
-
-  const roomType = g.r.GetType();
-  if (roomType !== RoomType.BOSS || !isRoomInsideGrid()) {
     return;
   }